* フォルダー整理 * ランチャー用のViewModelとViewを作成
