+++ DesktopCharacter/View/Menu/LauncherMenu.xaml
@@ -1,0 +1,26 @@
+﻿<MenuItem   x:Class="DesktopCharacter.View.Menu.LauncherMenu"
+            xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+            xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+            xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+            xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+            xmlns:local="clr-namespace:DesktopCharacter.View.Menu"
+            xmlns:vm="clr-namespace:DesktopCharacter.ViewModel.Menu"
+            xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity"    
+            mc:Ignorable="d" 
+            d:DesignHeight="300" d:DesignWidth="300"
+            Header="ランチャー" ItemsSource="{Binding Children}" >
+    <i:Interaction.Triggers>
+        <i:EventTrigger EventName="GotFocus">
+            <i:InvokeCommandAction Command="{Binding Focus}"/>
+        </i:EventTrigger>
+    </i:Interaction.Triggers>
+    <ItemsControl.DataContext>
+        <vm:LauncherMenuViewModel></vm:LauncherMenuViewModel>
+    </ItemsControl.DataContext>
+    <ItemsControl.ItemContainerStyle>
+        <Style TargetType="MenuItem">
+            <Setter Property="Header" Value="{Binding Headder}"/>
+            <Setter Property="Command" Value="{Binding Command}"/>
+        </Style>
+    </ItemsControl.ItemContainerStyle>
+</MenuItem>
