+++ DesktopCharacter/View/Setting.xaml
@@ -4,7 +4,7 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:DesktopCharacter.View"
-        xmlns:st="clr-namespace:DesktopCharacter.View.SettingTabView"
+        xmlns:st="clr-namespace:DesktopCharacter.View.SettingTab"
         xmlns:vm="clr-namespace:DesktopCharacter.ViewModel"
         mc:Ignorable="d"
         Title="Setting" Height="300" Width="300">
