+++ DesktopCharacter/View/SettingTab/TalkTab.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="DesktopCharacter.View.SettingTabView.TalkTab"
+﻿<UserControl x:Class="DesktopCharacter.View.SettingTab.TalkTab"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:DesktopCharacter.View.SettingTabView"
+             xmlns:local="clr-namespace:DesktopCharacter.View.SettingTab"
              mc:Ignorable="d" 
              d:DesignHeight="300" d:DesignWidth="300">
     <Grid>
